+++ JS/dog.js
@@ -3,6 +3,8 @@ const earthRadius = 48;
 
 // 创建一个对象来表示狗
 const dog = new THREE.Group();
+dog.isMoving = false; // Add a custom property to track movement
+dog.targetPosition = new THREE.Vector3(); // 目标位置
 
 // 创建身体
 const bodyGeometry = new THREE.BoxGeometry(2, 1, 1); // 宽度、高度、深度
